完成项目余下基本功能
- feat: 我的订单页面
- feat: 更新用户个人资料
- feat: 订单支付
- feat: 订单评论
- feat: 确认订单(填写收件人、手机及收货地址)
- feat: 删除未支付订单
- feat: 处理用户订单状态
- refactor: 将确定订单界面中的“删除订单”按钮移至我的订单页面
- fix: 将 Order 的时间类型从 Data 改成 LocalTime

+++ src/me/bqliang/servlet/AddCartServlet.kt
@@ -21,7 +21,7 @@ class AddCartServlet: HttpServlet() {
 
         val session = req.session
         // 构建购物车 item
-        val cartItem = ProductService.createCartItem(pid.toInt(), num)
+        val cartItem = ProductService.createCartItem(pid, num)
         // 从会话中获取购物车
         var cart: Cart? = session.getAttribute("cart") as Cart?
         // 如果购物车为空，则构建新的购物车对象
